# src/so101_teleop/so101_teleop/expert_validation_openapi.json
{
  "components": {
    "schemas": {
      "CampaignCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "command_id": {
            "minLength": 1,
            "title": "Command Id",
            "type": "string"
          },
          "lease_generation": {
            "minimum": 1.0,
            "title": "Lease Generation",
            "type": "integer"
          },
          "lease_id": {
            "minLength": 1,
            "title": "Lease Id",
            "type": "string"
          },
          "service_session_id": {
            "minLength": 1,
            "title": "Service Session Id",
            "type": "string"
          }
        },
        "required": [
          "service_session_id",
          "lease_id",
          "lease_generation",
          "command_id"
        ],
        "title": "CampaignCancelRequest",
        "type": "object"
      },
      "CampaignConfiguration": {
        "additionalProperties": false,
        "properties": {
          "execution_mode": {
            "enum": [
              "SEQUENTIAL",
              "PARALLEL",
              "ADAPTIVE"
            ],
            "title": "Execution Mode",
            "type": "string"
          },
          "fallback_worker_counts": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Worker Counts"
          },
          "initial_points_per_worker": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Points Per Worker"
          },
          "lease_generation": {
            "minimum": 1.0,
            "title": "Lease Generation",
            "type": "integer"
          },
          "lease_id": {
            "minLength": 1,
            "title": "Lease Id",
            "type": "string"
          },
          "manifest_id": {
            "minLength": 1,
            "title": "Manifest Id",
            "type": "string"
          },
          "max_infra_attempts_per_point": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Infra Attempts Per Point"
          },
          "max_points_per_worker": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Points Per Worker"
          },
          "preferred_worker_count": {
            "anyOf": [
              {
                "maximum": 16.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Worker Count"
          },
          "service_session_id": {
            "minLength": 1,
            "title": "Service Session Id",
            "type": "string"
          },
          "worker_count": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Count"
          },
          "worker_start_timeout_s": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Start Timeout S"
          },
          "yolo_executor_count": {
            "anyOf": [
              {
                "enum": [
                  1,
                  2,
                  4
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yolo Executor Count"
          }
        },
        "required": [
          "service_session_id",
          "lease_id",
          "lease_generation",
          "manifest_id",
          "execution_mode"
        ],
        "title": "CampaignConfiguration",
        "type": "object"
      },
      "CampaignStartRequest": {
        "additionalProperties": false,
        "properties": {
          "command_id": {
            "minLength": 1,
            "title": "Command Id",
            "type": "string"
          },
          "execution_mode": {
            "enum": [
              "SEQUENTIAL",
              "PARALLEL",
              "ADAPTIVE"
            ],
            "title": "Execution Mode",
            "type": "string"
          },
          "fallback_worker_counts": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Worker Counts"
          },
          "initial_points_per_worker": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Points Per Worker"
          },
          "lease_generation": {
            "minimum": 1.0,
            "title": "Lease Generation",
            "type": "integer"
          },
          "lease_id": {
            "minLength": 1,
            "title": "Lease Id",
            "type": "string"
          },
          "manifest_id": {
            "minLength": 1,
            "title": "Manifest Id",
            "type": "string"
          },
          "max_infra_attempts_per_point": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Infra Attempts Per Point"
          },
          "max_points_per_worker": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Points Per Worker"
          },
          "preferred_worker_count": {
            "anyOf": [
              {
                "maximum": 16.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Worker Count"
          },
          "preflight_receipt_id": {
            "minLength": 1,
            "title": "Preflight Receipt Id",
            "type": "string"
          },
          "service_session_id": {
            "minLength": 1,
            "title": "Service Session Id",
            "type": "string"
          },
          "worker_count": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Count"
          },
          "worker_start_timeout_s": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Start Timeout S"
          },
          "yolo_executor_count": {
            "anyOf": [
              {
                "enum": [
                  1,
                  2,
                  4
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yolo Executor Count"
          }
        },
        "required": [
          "service_session_id",
          "lease_id",
          "lease_generation",
          "manifest_id",
          "execution_mode",
          "command_id",
          "preflight_receipt_id"
        ],
        "title": "CampaignStartRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LeaseAcquireRequest": {
        "additionalProperties": false,
        "properties": {
          "service_session_id": {
            "minLength": 1,
            "title": "Service Session Id",
            "type": "string"
          }
        },
        "required": [
          "service_session_id"
        ],
        "title": "LeaseAcquireRequest",
        "type": "object"
      },
      "LeaseMutationRequest": {
        "additionalProperties": false,
        "properties": {
          "generation": {
            "minimum": 1.0,
            "title": "Generation",
            "type": "integer"
          },
          "service_session_id": {
            "minLength": 1,
            "title": "Service Session Id",
            "type": "string"
          }
        },
        "required": [
          "service_session_id",
          "generation"
        ],
        "title": "LeaseMutationRequest",
        "type": "object"
      },
      "ManifestCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "total_points": {
            "maximum": 20.0,
            "minimum": 4.0,
            "title": "Total Points",
            "type": "integer"
          }
        },
        "required": [
          "total_points"
        ],
        "title": "ManifestCreateRequest",
        "type": "object"
      },
      "RetryRequest": {
        "additionalProperties": false,
        "properties": {
          "command_id": {
            "minLength": 1,
            "title": "Command Id",
            "type": "string"
          },
          "confirmation": {
            "title": "Confirmation",
            "type": "string"
          },
          "lease_generation": {
            "minimum": 1.0,
            "title": "Lease Generation",
            "type": "integer"
          },
          "lease_id": {
            "minLength": 1,
            "title": "Lease Id",
            "type": "string"
          },
          "point_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Point Ids",
            "type": "array"
          },
          "service_session_id": {
            "minLength": 1,
            "title": "Service Session Id",
            "type": "string"
          }
        },
        "required": [
          "service_session_id",
          "lease_id",
          "lease_generation",
          "command_id",
          "point_ids",
          "confirmation"
        ],
        "title": "RetryRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SO-101 Expert Validation",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/expert-validation/artifacts/{artifact_id}": {
      "get": {
        "operationId": "artifact_expert_validation_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact"
      }
    },
    "/expert-validation/campaigns": {
      "get": {
        "operationId": "list_campaigns_expert_validation_campaigns_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Campaigns"
      },
      "post": {
        "operationId": "start_campaign_expert_validation_campaigns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Campaign"
      }
    },
    "/expert-validation/campaigns/preflight": {
      "post": {
        "operationId": "preflight_expert_validation_campaigns_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignConfiguration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preflight"
      }
    },
    "/expert-validation/campaigns/{campaign_id}": {
      "get": {
        "operationId": "get_campaign_expert_validation_campaigns__campaign_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Campaign"
      }
    },
    "/expert-validation/campaigns/{campaign_id}/cancel": {
      "post": {
        "operationId": "cancel_campaign_expert_validation_campaigns__campaign_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Campaign"
      }
    },
    "/expert-validation/campaigns/{campaign_id}/full-restart-retries": {
      "post": {
        "operationId": "retry_campaign_expert_validation_campaigns__campaign_id__full_restart_retries_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Campaign"
      }
    },
...
    "/expert-validation/lease": {
      "post": {
        "operationId": "acquire_lease_expert_validation_lease_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseAcquireRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acquire Lease"
      }
    },
    "/expert-validation/lease/{lease_id}": {
      "delete": {
        "operationId": "release_lease_expert_validation_lease__lease_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "lease_id",
            "required": true,
            "schema": {
              "title": "Lease Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseMutationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Lease"
      },
      "put": {
        "operationId": "renew_lease_expert_validation_lease__lease_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "lease_id",
            "required": true,
            "schema": {
              "title": "Lease Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseMutationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Renew Lease"
      }
    },
    "/expert-validation/manifests": {
      "post": {
        "operationId": "create_manifest_expert_validation_manifests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Manifest"
      }
    },
    "/expert-validation/manifests/{manifest_id}": {
      "get": {
        "operationId": "get_manifest_expert_validation_manifests__manifest_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "manifest_id",
            "required": true,
            "schema": {
              "title": "Manifest Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Manifest"
      }
    },
...
  }
}